music-ui: Widen last played and play count boxes
The last played is cut off for me in UK dates ("Fri 31 Aug 2018").

The right edge of the "Next" box (which is to the left of lastg
played) is at 200+600=800 so this leaves a buffer of 10px before
the new location at 810.

The right edge of the time box (which is to the left of the play
count) is at 715+140=855, so drop this to 715+125=840 to allow
10px before the new location of the play count at 850.