music-ui: adjust artist/album string wording
This is purely a preference thing, but I feel it makes more sense this way
round especially for compilation albums.

Taking for example the track:

    Various Artists - Terrorizer: Fear Candy 92/05-40 Watt Sun - Between Times.flac

Previously:

    Between Times
    From the album Terrorizer: Fear Candy 92 by the artist 40 Watt Sun

But the album "Terrorizer: Fear Candy 92" isn't by "40 Watt Sun", it's a
compilation by various.

Now it reads:

    Between Times
    By "40 Watt Sun" from the album "Terrorizer: Fear Candy 92"

Which I think is an improvement.